requirements.txt
jupyter
pandas
numpy
openai
pydantic
python-dotenv
matplotlib
tabulate
playwright
notion-client
